Documents/multi_domain_platform/requirements.txt
Python 3.13
bcrypt
pandas
Sqlite3
streamlit
google-generativeai